+++ CHANGELOG.md
@@ -1,12 +1,18 @@
+## 2.3.6 (under development)
+
+- Changed `davidhirtz\yii2\skeleton\modules\admin\Module::$name` to `Module::getName()` to prevent translation issues 
+  (Issue #28)
+
 ## 2.3.5 (Jun 28, 2024)
 
-- Fixed `AttributeTypecastBehavior` to cast integer "0" values to `0` not `null` 
+- Fixed `AttributeTypecastBehavior` to cast integer "0" values to `0` not `null`
 
 ## 2.3.4 (Jun 24, 2024)
 
 - Fixed MySQL JSON columns bug via migration. This normalizes JSON columns for MariaDB and MySQL with the introduction
   of JSON support in Yii 2.0.49.
-- 
+-
+
 ## 2.3.3 (Apr 24, 2024)
 
 - Extracted new `MessageController` actions to their own package `davidhirtz/yii2-translation`